+++ CHANGELOG.md
@@ -80,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Under the hood changes
 - merged from https://gitlab.com/distributed_lab/frontend/vue-template
 
-[Unreleased]: https://gitlab.com/tokend/nft-books/web-client-nft-books/compare/v1.2.0...main
+[Unreleased]: https://gitlab.com/tokend/nft-books/web-client-nft-books/compare/v1.3.0-rc.1...main
+[1.3.0-rc.1]: https://gitlab.com/tokend/nft-books/web-client-nft-books/compare/v1.2.0...v1.3.0-rc.1
 [1.2.0]: https://gitlab.com/tokend/nft-books/web-client-nft-books/compare/v1.1.0...v1.2.0
 [1.1.0]: https://gitlab.com/tokend/nft-books/web-client-nft-books/compare/v1.0.0...v1.1.0
 [1.0.0]: https://gitlab.com/tokend/nft-books/web-client-nft-books/tags/v1.0.0